requirements.txt
# Pinned for macOS/Python 3.12 compatibility
numpy<2
torch==2.2.2
sentence-transformers<3.0
transformers<4.42

requests>=2.31.0
markitdown[docx]>=0.0.1
html-to-markdown>=0.1.0
pypdf>=4.0.0
langchain-text-splitters>=0.0.1
pydantic>=2.5.0
click>=8.1.0
pytest>=7.4.0
pytest-mock>=3.12.0
pyyaml>=6.0.1
google-generativeai>=0.8.0
google-genai
qdrant-client>=1.7.0
newspaper3k>=0.2.8
cloudscraper>=1.2.71
lxml_html_clean>=0.4.0
markdown>=3.8.0
ruff>=0.12.0
